pyproject.toml
[project]
name = "azure-ai-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-evaluation>=1.16.6",
    "azure-core>=1.39.0",
    "azure-identity>=1.25.3",
    "pip>=26.1",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=4.6.0",
    "pyright>=1.1.409",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.12",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.report]
fail_under = 90
exclude_also = [
    'if __name__ == .__main__.:',
]


